+++ src/utils.js
@@ -1,3 +1,5 @@
+const fs = require("fs");
+
 /**
  * Generates a unique ID.
  * @returns {string} The unique ID.
@@ -16,6 +18,11 @@ function generateUniqueId() {
   return objectId.substring(0, 24);
 }
 
+/**
+ * Logs message to the console, with an optional type parameter to specify the log level.
+ * @param msg - Represents the message to log.
+ * @param type - Specifies the type of message. It can be either "error" or any other value.
+ */
 function logger(msg, type) {
   if (type === "error") {
     console.error(msg);
@@ -24,4 +31,14 @@ function logger(msg, type) {
   }
 }
 
-module.exports = { generateUniqueId, logger };
+/**
+ * Checks if a directory exists, and if not, creates it.
+ * @param directoryPath - String that represents the path of the directory
+ */
+function dirCheck(directoryPath) {
+  if (!fs.existsSync(directoryPath)) {
+    fs.mkdirSync(directoryPath, { recursive: true });
+  }
+}
+
+module.exports = { dirCheck, generateUniqueId, logger };